requirements.txt
asgi-redis==1.4.3
asgiref==3.4.1
astroid==2.7.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.4
cloudinary==1.26.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
daphne==3.0.2
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-redis==5.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack-python==0.5.6
openapi-codec==1.3.2
paramiko==2.7.2
Pillow==8.3.1
platformdirs==2.2.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pylint==2.10.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pysftp==0.2.9
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
service-identity==21.1.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
